% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{apply_transforms}
\alias{apply_transforms}
\title{Manually apply transforms to imgs}
\usage{
apply_transforms(wb)
}
\arguments{
\item{wb}{A \code{wb} object}
}
\description{
While functions like \code{wb_present} will automatically apply transformations,
this function allows you to apply them manually.
}
