% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_annot.R
\name{col_annot}
\alias{col_annot}
\alias{col_annot<-}
\title{Get or set column annotation for a wb object}
\usage{
col_annot(x)

col_annot(x) <- value
}
\arguments{
\item{x}{A \code{wb} object}

\item{value}{A data.frame with each column representing a variable, and each
row representing a lane in the blot. NOTE: the order of the rows should
annotate the lanes left-to-right AFTER image manipulation, which might
include mirroring.}
}
\description{
Get or set column annotation for a wb object
}
