% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgs.R
\name{imgs}
\alias{imgs}
\alias{imgs<-}
\alias{wb_add_img}
\title{Get or set images for a \code{wb} object}
\usage{
imgs(x)

imgs(x) <- value

wb_add_img(x, imgs, names)
}
\arguments{
\item{x}{A \code{wb} object}

\item{value}{Unused.}

\item{imgs}{A \code{magick-image} containing one or more images}

\item{names}{A character vector of names with length equal to \code{imgs}. Can be
\code{NULL} if row_annot is \code{NULL}}
}
\description{
Get or set images for a \code{wb} object
}
\details{
Note that \verb{imgs<-} is unused - add images using \code{wb_add_img}
}
