% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb.R
\name{is_wb}
\alias{is_wb}
\title{Test if the object is a wb}
\usage{
is_wb(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
Logical. \code{TRUE} if object inherits from the \code{wb} class
}
\description{
Test if the object is a wb
}
