% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\name{present_wb}
\alias{present_wb}
\title{Render a wb object into an image}
\usage{
present_wb(wb)
}
\arguments{
\item{wb}{A \code{wb} object with at least images and transformations}
}
\description{
This function takes a wb object and (optionally) row and column annotations
and produces an image from them.
}
