% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_annot.R
\name{row_annot}
\alias{row_annot}
\alias{row_annot<-}
\title{Get or set row annotation for a wb object}
\usage{
row_annot(x)

row_annot(x) <- value
}
\arguments{
\item{x}{A \code{wb} object}

\item{value}{Either a \code{data.frame} with a single column \code{name}, or a
character vector}
}
\description{
Get or set row annotation for a wb object
}
