% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transforms}
\alias{transforms}
\alias{transforms<-}
\title{Get or set transforms of a \code{wb} object}
\usage{
transforms(x)

transforms(x) <- value
}
\arguments{
\item{x}{A \code{wb} object}

\item{value}{A \code{data.frame} with the same number of rows as \code{row_annot}, if
not null, and with columns \code{width}, \code{height}, \code{xpos}, \code{ypos}, \code{rotate}, and
\code{flip}}
}
\description{
Get or set transforms of a \code{wb} object
}
