% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgs.R
\name{wb_add_img}
\alias{wb_add_img}
\title{Add an image to a wb object}
\usage{
wb_add_img(x, imgs, names)
}
\arguments{
\item{imgs}{A \code{magick-image} containing one or more images}

\item{names}{A character vector of names with lenght equal to imgs}
}
\description{
Add an image to a wb object
}
