% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgs.R
\name{wb_rm_img}
\alias{wb_rm_img}
\title{Drop an image and its band data by name or index}
\usage{
wb_rm_img(x, name)
}
\arguments{
\item{x}{A \code{wb} object}

\item{name}{Either a numeric representing the index of the image, or a name
identifying the row_annot row associated with the image.}
}
\description{
Drop an image and its band data by name or index
}
