% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_visual_edit.R
\name{wb_visual_edit}
\alias{wb_visual_edit}
\title{Interactively edit blots}
\usage{
wb_visual_edit(wb)
}
\arguments{
\item{wb}{A \code{wb} object}
}
\description{
This function helps generate code needed to transform a blot to your
specifications via a Shiny app
}
