% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgs.R
\name{imgs}
\alias{imgs}
\alias{wb_add_img}
\title{Get and set images from a wb object}
\usage{
imgs(x)

wb_add_img(x, imgs, names)
}
\arguments{
\item{x}{A \code{wb} object}

\item{imgs}{A \code{magick-image} containing one or more images}

\item{names}{A character vector of names with length equal to \code{imgs}. Can be
\code{NULL} if row_annot is \code{NULL}}
}
\description{
Get and set images from a wb object
}
